Move watcherapi controller to controllers directory
For consistency with other operators and with defaults in operators-sdk.
